FBA Refund Predictor/mcde/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML and data processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2

# OCR and document processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
opencv-python==4.8.1.78

# Computer vision and image processing
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2

# Database and storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0

# API and HTTP
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1

# Configuration and logging
pyyaml==6.0.1
loguru==0.7.2
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.11.0
mypy==1.7.1

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Development tools
pre-commit==3.6.0

# Detection Engine Dependencies
asyncio-mqtt==0.16.1 